Switch substitute store to exclusive scope cache
This is important for tests that use store substitutes. Without this
commit, each time a substitute store is constructed, the same actual
cache is used. So if one test inserts an object into the cache for a
given store scoped to that test, all other stores in all other tests
will still use that same cache.